remove material and related from mkdocs
These are a bit high maintenance for me. I also removed the kinds of
tracker/"social" stuff for the project, sinc most projects aren't going
to need a facebook page and twitter or others avenues of communication.

In general, having fewer pipelines of information to monitor for a
project will reduce reach, but make it easier on the maintainer.